remove jni call that only removes one layer of encryption from encrypted query results
